Create README.md
Update for WhatsApp updates to the latest version for WhatsApp from Official Website directly. You can install Update for WhatsApp to your any android tablets and smartphones.
As you may know, new version for WhatsApp is published more often on the official Website than in the store. By using this application you can check whether new version is available and update it. This application also periodically notifies you whenever new version for WhatsApp is available.

Main Features :
*  Easy to Use
* Notifies Periodically Whenever New Version for WhatsApp is Available
* Using New Features for WhatsApp before your friends.
* Admob Integrated
* Ready to publish


Important:
WhatsApp name is copyright to WhatsApp,Inc. This update for WhatsApp is in no way affiliated with, sponsored or endorsed by WhatsApp, Inc. We are not responsible for any kind of re-usage of any whatsapp status downloaded by the user.